[sigverify] Update SPHINCS+ to use aligned signatures and keys.
Since ROM public keys and signatures are already required to be
word-aligned, we can align a bunch of internal buffers in SPHINCS+ and
remove a cast-to-bytes for sigverify. The performance impact is
basically nil (about 1000 cycles faster out of a total ~1M cycles),
because I think the compiler already noticed that most of these were
aligned.

+++ sw/device/silicon_creator/lib/sigverify/sphincsplus/utils.h
@@ -49,9 +49,9 @@ uint64_t spx_utils_bytes_to_u64(const uint8_t *in, size_t inlen);
  * @return Error code indicating if the operation succeeded.
  */
 OT_WARN_UNUSED_RESULT
-rom_error_t spx_utils_compute_root(const uint8_t *leaf, uint32_t leaf_idx,
+rom_error_t spx_utils_compute_root(const uint32_t *leaf, uint32_t leaf_idx,
                                    uint32_t idx_offset,
-                                   const uint8_t *auth_path,
+                                   const uint32_t *auth_path,
                                    uint8_t tree_height, const spx_ctx_t *ctx,
                                    spx_addr_t *addr, uint32_t *root);
